Added mouse interaction with the piano roll to add/remove notes

+++ src/main/ui/common/LineContainerPanel.java
@@ -29,7 +29,7 @@ public abstract class LineContainerPanel extends JPanel {
     // MODIFIES: this
     // EFFECTS: updates lineX value with the timeline tickPosition (scaled) and repaints
     public void updateLineX() {
-        this.lineX = (int) timelineController.getTimeline().scaleTickToPixel(player.getTickPosition());
+        this.lineX = timelineController.getTimeline().scaleTickToPixel(player.getTickPosition());
         repaint();
     }
